Moving styles to separate files
