+++ src/components/navigation-header/components/home-page-content/home-page-content.module.css
@@ -1,0 +1,3 @@
+.siteLogo {
+  display: flex;
+}
