+++ src/components/navigation-header/components/primary-nav-link/primary-nav-link.module.css
@@ -1,0 +1,17 @@
+.root {
+  /* Composition */
+  composes: g-focus-ring-from-box-shadow-dark from global;
+
+  /* CSS properties */
+  border-radius: 5px;
+  color: var(--token-color-palette-neutral-400);
+  padding-bottom: 8px;
+  padding-left: var(--header-dropdown-menu-padding-left-right);
+  padding-right: var(--header-dropdown-menu-padding-left-right);
+  padding-top: 8px;
+
+  &:hover {
+    color: var(--token-color-surface-primary);
+    cursor: pointer;
+  }
+}
