+++ src/components/navigation-header/components/product-page-content/product-page-content.module.css
@@ -1,0 +1,22 @@
+.companyLogo {
+  display: flex;
+
+  & svg {
+    height: 32px;
+    width: 32px;
+  }
+}
+
+.companyLogoMenuButton {
+  padding-bottom: 2px;
+  padding-top: 2px;
+}
+
+.productLogoLink {
+  /* Composition */
+  composes: g-focus-ring-from-box-shadow-dark from global;
+
+  /* CSS properties */
+  border-radius: 5px;
+  margin-left: 12px;
+}
