+++ src/components/navigation-header/navigation-header.module.css
@@ -35,34 +35,13 @@
 * off-center vertically.
 ***
 */
-.siteLogo,
-.companyLogo,
 .productLogo {
   display: flex;
-}
 
-.companyLogoMenuButton {
-  padding-bottom: 2px;
-  padding-top: 2px;
-}
-
-.companyLogo svg {
-  height: 32px;
-  width: 32px;
-}
-
-.productLogoLink {
-  /* Composition */
-  composes: g-focus-ring-from-box-shadow-dark from global;
-
-  /* CSS properties */
-  border-radius: 5px;
-  margin-left: 12px;
-}
-
-.productLogo svg {
+& svg {
   height: 40px;
   width: auto;
+}
 }
 
 .rightSide {
@@ -97,22 +76,4 @@
   list-style: none;
   margin: 0;
   padding: 0;
-}
-
-.mainNavLink {
-  /* Composition */
-  composes: g-focus-ring-from-box-shadow-dark from global;
-
-  /* CSS properties */
-  border-radius: 5px;
-  color: var(--token-color-palette-neutral-400);
-  padding-bottom: 8px;
-  padding-left: var(--header-dropdown-menu-padding-left-right);
-  padding-right: var(--header-dropdown-menu-padding-left-right);
-  padding-top: 8px;
-
-  &:hover {
-    color: var(--token-color-surface-primary);
-    cursor: pointer;
-  }
 }
